refactor(ui): split keys.go by feature
keys.go keeps root key routing, cursor movement, reorder and
collapse. Session lifecycle (attach/revive/kill/archive/delete),
rename, quick prompt, settings, and move each get their own file;
focus-mode handlers join focuskeys.go and openDiff joins
diffview.go. Pure moves.
